Expose WordPressUtils for consumption by gutenberg-bridge
WPAndroid updating to Gradle 5.4.1 caused a build failure when
BUILD_GUTENBERG_FROM_SOURCE was set to true because the
react-native-gutenberg-bridge could not access the WordPressUtils. In
particular, the WPAndroidGlueCode class would fail to import
org.wordpress.android.util.AppLog, causing a compilation error.
Switching WordPressUtils from an implementation dependency to an api
dependency in react-native-aztec resolves this problem.