Build JS bundle and upload AAR to Bintray in one step on CI
So far, we've been unable to run the JS bundle build from within the
Gradle task because of memory constraints in the Docker image that
CircleCI provides — granted, we're on the medium one, not the large.

By switching over to a bare bone Linux machine and taking care of
setting up Android ourselves, we gained better memory and were able to
finish the build. Yey!

This started from the conversation in
https://github.com/wordpress-mobile/gutenberg-mobile/pull/3038#discussion_r568136000,
which also explains the benefits of relying on Gradle to do everything.