Fix CI shallow git clone
Previously, the CI configuration cloned the repository with a depth of
1. In the context where the CI job was attempting to check out an older
commit of a branch, this would cause an error. This might occur for a
scheduled job on a shared branch when a merge occurs after the scheduled
job begins. This commit increases the depth to reduce the likelihood of
this circumstance.