CI fetches commit reference directly
The CI configuration is set to clone with a depth of 1. If the CI tasks
then attempts to reference an older commit in the relevant branch, an
error occurs. This might occur when a branch merge occurs between a CI
tasks beginning and the point it attempt to fetch the HEAD of the
branch. This change configures the shallow clone to fetch the specific
commit reference, rather than the HEAD of the relevant branch.