Update the GH workflow event listened for from `on.push.tags` to `on.release.published` as it would be more accurate.
Presumably, we don't need create a WPA PR every time a tag is published. If you are actively working on changes, then you would create a PR for GBM and use the `sync:android` label or `npm run sync:android -- --pr NUMBER`. However, if you are getting ready to release changes, presumably you would open a GBM PR from a `release/` branch targeting `trunk`, and if everything looks good you would merge the branch and then create a `tag` and update WPA to reference the new GBM `tag`. So there isn't inherent value in creating a WPA PR every time a `tag` is pushed as every case should be covered adequately with a GBM PR. There is value in having a WPA PR created implicitly from the promotion of the newly created GBM `tag` to a `release` so that WPA can reference the new GBM `release` right when it is created.

This is what is attemoting to be addressed here, and was discussed thoroughly yesterday.