Update Jeptack shared extension util requirement
`@automattic/jetpack-shared-extension-utils` uses a dev requirement that
is required from a workspace using pnpm syntax (npm@7 has workspace
support but the sytnax is different). That dependency fails when trying to
install the node packages from Gutenberg Mobile. Setting the
dependency in Gutenberg Mobile as optional will ignore the sub dependency
faliure, which is not ideal but since the sub dependency is not needed it's
ok.