Fix logic so that VideoPress block is dev-only
The VideoPress block was previously displaying in all builds, as the __DEV__ flag had been incorrectly added to the 'supportedJetpackBlocks' object. As the block belongs in a different packages to 'regular' Jetpack blocks, it was not passed to the 'register-jetpack-block.native.js' file, so didn't go through the usual availability checks.

As a way around this, the __DEV__ flag has been added alongside the 'hideBlockByCapability' function. The unnecessary inclusion of VideoPress in the 'supportedJetpackBlocks' object has been removed.