test: Fix Android block appender button query
The available elements within this context are unexpected, with certain
elements unavailable or covered by non-visual elements, e.g. resize
observer. I consider this change a workaround. There is likely a better
way of achieving this, possibly an existing helper that could be used or
modified to support this context.