Add `bundle/ios/App.js` to the resources bundles with the XCFramework
The React Native iOS runtime needs this file, albeit converted as a
`jsbundle`.

The Jetpack and WordPress iOS setup currently use `App.js` and converts
it to a `jsbundle` file in a build phase run script. For the moment
we'll stick with that approach, looking at possible refinements later
on.