HACK – Port all XCFramework-publishing automation from `trunk` on 1.97.1
This is just so we can fetch the XCFramework for 1.97.1 in the Jetpack
and WordPress iOS apps, even though there is not an official tag build
for it.

The practical benefit is that we can merge the integration work without
having to wait for a new Gutenberg mobile version to ship.