test: Add to-do note regarding test improvements
In general, we should not wrap `fireEvent` in an `act` call, as it
already leverages `act` itself. Wrapping it is an indicator of a deeper
issue: there are state updates for which we are not awaiting. Ideally,
we add a new query to await the addition/update/removal of UI to satisfy
the act warning.