test: Remove default Jest log verbosity
The current approach of enabling the `verbose` option for all test
scripts is counter to the Jest default and the Gutenberg web editor
preset. This better aligns the native tests with the web tests.

The `verbose` option remains enabled for "debug" scripts and can
be enabled ad-hoc via the `--verbose` flag.