Inline .podspec.json extension definition
I noticed this because RuboCop barked at me for having a method with 11
SLOC instead of 10.

However, I'm happy with the change. The implementation expects the file
to be a JSON, so there's little value in making the extension
configurable. If we changed that var to `.podspec`, the implementation
would break!