build: Revert Aztec bump
The gutenberg branch associated with this PR is behind gutenberg
`trunk`, which includes a bump of Aztec. Therefore, we need to revert
the bump in gutenberg-mobile temporarily to allow the CI tasks to
succeed.

Whenever this branch is merged, we should prefer the latest Aztec
version, which is already in gutenberg and gutenberg-mobile the `trunk`
branches.