build: Correct lock file changes
The previous lock file was incorrect, as it was the result of running
`npm install` after merely bumping package version numbers. In order to
capture the correct lock file results, one must follow the Gutenberg
project's guidelines for updating dependencies, which involves removing
and re-adding each dependency. This now captures the result of `npm
install` after following that process.

https://github.com/WordPress/gutenberg/tree/c7cf2f74f38ae4c1f325e3238e61270a801f183c/packages#updating-existing-dependencies