build: Add react-native dev dependency
Some CLI commands were hoisted from the dedicated CLI package into
React Native core. Without React Native installed in the top level of
this project, scripts will fail due to missing commands.

This is considered a workaround currently. It would be ideal to avoid
the need to duplicate this dependency that is already within the
Gutenberg core project, but it may be unavoidable due to the CLI changes
in core.